Change readme.md to state diag. tear. on hardware
The diagonal line is an artifact of GraphX using a row-major ordering of
pixels in the VRAM instead of column-major ordering. The screen is
actually a portrait display, requiring the hardware to flip the screen
over the diagonal line, causing minor screen tearing. The only way to
fix this is to rewrite GraphX to use the column-major ordering that is
native to the display. You can view the same effect using Ndless
programs on the TI-Nspire CX II (CAS) because that calculator uses the
same display.